attempting to split up actions to reduce branching

+++ bwi_guidance_solver/include/bwi_guidance_solver/person_model_iros14.h
@@ -76,7 +76,7 @@ namespace bwi_guidance {
       /* Helper Functions */
       float getTrueDistanceTo(const RobotStateIROS14& state, int destination);
       int selectBestRobotForTask(int destination, float time_to_destination);
-      bool isRobotDirectionAvailable(float& robot_dir);
+      bool isRobotDirectionAvailable(int& robot_dir);
       int generateNewGoalFrom(int idx);
 
       /* Action generation caching */
